+++ R/vst.R
@@ -4,6 +4,8 @@
 #' @export
 
 vst_ps_to_mx <- function(ps) {
+  .check_suggests(c("DESeq2", "SummarizedExperiment"),
+                  "for the variance-stabilizing transformation.")
   phyloseq_to_deseq2(ps, ~ 1) %>% # DESeq2 object
     DESeq2::estimateSizeFactors(., geoMeans = apply(
       DESeq2::counts(.), 1, function(x) exp(sum(log( x[x>0] )) / length(x)))) %>%